[02] Error handling
Taking a more Rust-y solution for this step by taking advantage of the
possibility of returning a `Result` from the main function. Like for `perror()`
and `errno` for C, this allows the program to return error with a code and a
description. This way we can fully use the Termios API and directly return the
errors encountered when something goes wrong in `enable_raw_mode()` or
`disable_raw_mode()`. The following example show that we get the same error as
we would with the C implementation (with some additional Rust formatting).

```bash
$ echo test | ./target/debug/kg
Error: Os { code: 25, kind: Other, message: "Inappropriate ioctl for device" }
```