[03] Refactor keyboard input - part 1
Extract character reading from stdin and move it to `editor_read_key()`. Instead
of just returning a char and having the potential side effect of exiting the
program (like the C implementation), we wrap the return value in a
Result (either the char or an error).

The result code is a bit more verbose than the original implementation but the
error handling is made explicit.

Switched to using `u8` instead of `char`:
- no need for cast `as char` everytime a `read()` is made
- `b''` shortcut makes it easy to define char in the code