[03] Clear the screen on exit
Skipping the `atexit()` pattern makes it hard to clean the screen when an error
is thrown. One possibility would be to wrap the code into a sub-function and not
throw from the `main()`. An other possibility would be to piggyback on a
destructor of a struct initialized in the main.