[03] Global state
By defining a `drop()` method on the structure, we can implement the equivalent
of the `atexit()` from the C implementation. The function will be triggered when
the structure is deallocated, at the end of the scope of the main function. To
make it more symmetrical we can also use the `new()` method to set up the raw
mode.

All the `editor_*` functions now all take the EditorConfig as first argument.
That makes their signature more homogeneous and should make up for not storing
the editor config as a global variable.

Rust specific consideration: We could host all `editor_*` functions under the
EditorConfig name and even a dedicated `editor` module. For now we will stick to
the format of the tutorial so it's easier to follow.