[03] Window size, the easy way
Instead of returning `0` or `-1` to indicate success or error and pass the
values through references we can use Rust's Result type.

Once again, instead setting the global variables we create the instances
locally. This means that the `init_editor()` function is a bit bigger than the
original C version.

Renamed the `Error` variants for more idiomatic naming. For reference see:
https://rust-lang.github.io/rust-clippy/master/index.html#enum_variant_names