[03] Window size, the hard way - part 1
The Simple Error construct will be reused quite often so we can use this step as
an excuse to refactor it a bit. Given the use of the `line` and `column` macros,
we cannot use a function to return the `Err`, otherwise the errors would all
have the same line/column information.

We will print and read escape sequences from the terminal. To not have the
sequences printed to the terminal, the raw mode needs to be enabled before we
get the window size.