[03] Welcome message - part 2
Centers the welcome message. Use `vec!` macro to generate all the space needed
instead of looping.

`cargo clippy` suggests using explicit match instead of a comparision chain to
make sure the cases are exhaustive. It does have the advantage of making the
intention explicit, although for this particular case it is a bit verbose.