[03] Move the cursor - part 2
To be able to use `match` on the keys inline values, enums or const needs to be
provided. That's why `ctrl_key()` was converted to `const` and `CTRL_Q` is a constant.

To make it a bit more clean when using the u8 key values, everything was moved
to a private sub module. All the values defined here will be available by using
`key::*`.

All cursor move events will be represented as a single enum value containing the
necessary information to update the cursor. An enhancement would be to use an
enum to encode the forward / backward / upward / downward movements to not tie
them to key values (useful for supporting different keyboard layout for
example).