[03] Arrow keys - part 1
Here we can use advanced pattern matching features to to match directly against
the content of the array and not have to test the chars one by one.

As the escape sequences we expect to handle are all 2 bytes long, the `seq`
buffer is of length 2. We will update it when we need to handle longer escape
sequences.