[03] Arrow keys - part 2
This step works well in Rust but starting showing the limits of the 1:1
translation from the C implementation. In the C implementation the enum is set
up to prepare for a bigger refactor (in the next step) which will (ab)use the
fact that enum can be assigned `int` values automatically and that `char` is a
subset of `int`.

To have the same effect the Rust implementation needs to use constants instead
of enums and the value will need to be defined manually. At some point, it was
possible to "convert" Rust enums to C-like enums through the `FromPrimitive`
trait. This trait has however been deprecated and removed a later versions of
Rust. One now needs to use macros to achieve this particular conversion.

Given that we don't actually need to interface with C code, we should diverge
from the original implementation and make use of the full power of Rust's enums.