[03] Arrow keys - part 3
In contrast with the C implementation where special keys are part of the
`editorKey` enum and default chars are represented as chars, the Rust enums can hold different value
types and allow to represent all keys within a single enum.

This allows for the `editor_move_cursor()` to exhaustivly match on all possible
arrow direction and have that constraint checked by the compiler.

It also allows for a better match pattern in `editor_process_keypress()` where
we can forward the arrow direction and delegate the exhaustivness check to
`editor_move_cursor()` instead of listing all cases in both functions.