[03] Arrow keys - part 4
Extra step to clean up the Rust implementation after having caught up with it in
the last step.

The Ctrl keys are now represented as such within the enum. Instead of computing
each possible Ctrl keys and storing them as constant we can match any Ctrl key
through the `Key::Ctrl(char)` enum variant. This makes the support for Ctrl-Q
fall in with the same pattern as the arrow keys.