[03] The Page Up and Page Down keys - part 2
Piggybacking on direction enum that was initially defined for the arrow keys
made it trivial to implement the Page Up / Page Down keys. Augmenting the
"cursor move" event to support a number of moves allows a nice extension of the
`editor_move_cursor()` function to support this new movement without duplicating
code mutating the cursor position.

One problem with the Rust implementation is that we need to be mindful of the
integer overflow when doing additions and subtractions. If left unchecked
overflows will trigger panics. Using saturating add/sub the unsigned integer
will stay between zero and max integer. Different overflow behaviours can be
achieved with other functions for controlled overflow.