[04] A line viewer - part 1
Once again, we can rely on the standard `Vec` type to manage memory allocation
automatically. This makes this part a smaller changeset than its C counterpart.

Using `Vec<String>` as the type for the rows makes it easier for this first part
but it might be more advantageous to switch to a `Vec<Vec<u8>>` representation
to not have to convert to bytes before adding the text to the "append buffer".