[04] Multiple lines
Fixed the display of the last line which was simplified in the `Window size, the
easy way` section.

Skipped a few changes in this part as well. Using `Vec` and array still make
things much easier than the C implementation. If any later parts rely on this
changes it will be a good excuse to try Rust's refactoring capabilites.