[04] Vertical scrolling - part 2
Interesting fact: Need to use `cy + 1 - row_offset` over `cy + 1 - row_offset`
otherwise the system will panic when `cy = 0` and `row_offest = 1` because `cy`
is unsigned.

The `u16` vs `usize` types is starting to show up in more and more function.
Converting everything to `usize` might make the code cleaner.

The implementation currently follows the C version and has the drawback of
pushing `editor_refresh_screen` to mutate the global state. It might be cleaner
to isolate mutations due to scrolling with the cursor movement in the
`editor_move_cursor` function. That way the function `editor_refresh_screen`,
which handles the construction of the UI, would remains pure. This might bring
benefits in the long term if we would want to run it in another thread for
example.

Arguably the scrolling should be made as an event which could be processed by
the `editor_move_cursor` function. However the system currently only handles one
event per iteration. We would need to process all events before triggering a new
`editor_process_keypress`.