[04] Horizontal scrolling - part 4
Make sure the cursor cannot go further than the end of the line.

Given the earlier choice of handling a number of "times" for the cursor to move
in the `editor_move_cursor` function, we need to handle cases where `times > 1`.
This makes the change a bit more verbose than the C version.