[04] Snap cursor to end of line
The cursor is now moved to the end of line when moving up/down to a line smaller
than the current cursor position. It doesn't move it back to the original
position when going back to the previous line. This could be done by saving the
previous max position of the cursor.

The `usize` vs `u16` makes for some potentially panic-inducing conversion at
runtime.