[05] Insert ordinary characters
Adds a new `InsertChar` event which is created when a "char" key is pressed.
This event currently both insert the character and moves the cursor to the
right.

Updates `Key::Char` to hold a `char` instead of a `u8` which will better match
the semantic and the usage of `Char`.

Contrary to the original implementation, and because we decided previously not
to implement an explicit `editor_append_row`, `editor_insert_char` needs to
create a new ERow when adding a new line. Given that the row will only contain
one character, it is formatted during creation and does not rely on
`editor_row_insert_char`. Still, `editor_row_insert_char` only knows how to
update an ERow.

In the original version, `editor_insert_char` updates the cursor position `E.cx`
and relies on `editor_refresh_screen` to update the rendered position. In this
implementation `editor_refresh_screen` does not mutate state, thus an explicit
call to `editor_move_cursor` is needed to achieve the same behaviour.