[05] Save to disk - part 1
Instead of just returning a pointer to the String created in
`editor_rows_to_string` we need to give the ownership of the String to the
caller. The compiler requires it to ensure thhat the caller will free the
memory (in our case it will be automatically freed at the end of the block in
`editor_save`). Contrary to the original C version, it's impossible to forget to
free the string.