[05] Dirty flag
Used a boolean instead of an integer as the value will only be used as a
boolean. This saves us from having to handle potential integer overflow.

Updated the signature of the `editor_save` function. Now that the EditorConfig
needs to be mutable to set the dirty flag, we can move part of the logging back
in the function.