[05] Save as... - part 1
To be able to match the `filename` type when checking for its presence in
`editor_save`, the `new_filename` variable needs to be defined outside of the
if/else block. However it can be initialised only in the else block where it's
needed. The compiler will check that all use of the variable will only be done
after initialisation.