[06] Search - part 2
Added more tests around conversion of cursors.

Do not tests situation were the cursor is moved from another line. The current
behaviour is incoherent but fixing it is not the current priority.

If we were to move the tests to another file, most of the functions would also
have to be moved out of `main.rs`. Rust does not allow module import from the
`main.rs` file.