[06] Search forward and backward - part 2
Interesting Rust trivia: `iter().filter().next()` is considered equivalent to
`iter().find()` when it comes to returning the first filtered value. The latter
form is preferred as it short-circuits (the filter will not be applied to all
the elements iterated on). `clippy` also suggest replacing the former with the
latter. However it must be noted that the `filter().next()` combo works on both
immutable and mutable iterators while `find()` requires mutable iterrators.

Search can be improved on multiple points:
- Up and Down arrows should not retrigger a full swipe of the rows.
- Up and Down arrows should use the fact the the matches will be returned as a
  set of sorted cursor positions. Iterating over all matches is not needed, only
  the closet matches (previous and next) need to be considered.
- Adding a new letter to the query should trigger a search through the matches
  that have already been found instead of searching again through all the lines.