[07] Restore syntax highlighting after search
Changed the `editor_prompt` callback type from `fn` (type of a compiled
function) to the `Fn` trait to allow the use of closures. That way we can "save"
the document highlighting and restore it between each iteration of the search.
This is to replace the use of the static variables in the C implementation.

The major caveat is that closure are allocated at runtime and thus cannot be
used with container types such as Option or Result with a Box. This would lead
to having all the closure allocated on the heap which is not optimal. So
instead, the default value for the callback paramter from `editor_prompt` will
be an empty closure (as opposed to passing Option::None previously).

Given how sparse the search hilghlights will be compared to the default search,
it should be possible to store the information in a lighter way than by cloning
the whole `hl` matrix. Saving the matches position in the editor's state would
be an option.