[07] Detect filetype - part 1
The `const` type of HLDB makes for some interesting definition of the
`EditorSyntax` struct. Particularly the `file_match` array currently as a size
of 1 which means that every and any new EditorSyntax would only be allowed one
file extension, which defeats the purpose of having an array. This will have to
be fixed once the editor has support for multiple file types.