[07] Detect filetype - part 7
The last refactoring for this is part is the use of a reference in the signature
of `HLDB`. As a `const` the value of `HLDB` will be copied where ever it is
referenced in the code. Using a reference may be preferable depending the size
of the array. Given the only use of the HLDB (iterating over it), the use of a
reference fits better than copying it and creating an iterator at runtime.