[07] Colorful keywords - part 2
- Split the different highlight cases and turn each one of them in dedicated
functions.
- Pass the string around as a [char] instead of using an iterator to make it
easier to look ahead/behind without consuming the iterator and having to pass
chars around when we're in the wrong case (e.g. single line comments). This was
also a limitation for implementing the keyword highlighting.

In a best of world, we'd have a dedicated parsing system that would build the
chunks to highlight directly and we wouldn't have to iterate char by char. To
keep as close as possible to the original implementation, this will have to do
for now.