[07] Colorful keywords - part 7
Refactored the should_highlight_* highlight_* functions to only process the data
once when checking for keywords. Instead of using a conditional clause for each
case, the highlight functions are now chained: the first one to return an
`Option::Some` will be used to update the value of the cursor and the remaining
highlight functions will not be processed.