[07] Colorful multiline comments - part 3
This change is a refactor on the use of `ERow`. While the structure has changed
the code stays functionally equivalent.

1. Update `editor_update_syntax` to work on a mutable `ERow` instead of just the
"render" part. Now the function can:

  1.1 Work on the "hl" in place instead of returning it for the caller to update
  the row with the result

  1.2 Make it possible to return extra information about the state of the
  highlight, which will be needed for the multiline comments

2. Adds two new fields to the `Erow` structure (`hl_open_comment` and
`hl_close_comment`). These fields are not used in the implementation but will
require all uses of instantion of `ERow` to be updated, thus making it easier to
track the usage with the help of the compiler.

3. To replacce all manual instanciations of the `ERow` structure, a new
`ERow::from` function was added to create Erows from a string. The render and
highlight will be computed during creation. This limits repetition and make it
easier to update the code for the next change of the `editor_update_syntax`
function.

4. Following 3, all assertions of the `editor_udpate_syntax` tests have been
updated. In this case we use a macro to simplify the syntax of assertions and
keep the tests as clear as possible.