Add support for reading Unicode characters - part 2
Update ERow's `render` field to match representation of representation `chars`.
Without it the "cursor" and the "render cursor" cannot be kept in sync.

While it now supports all Unicode characters and the cursor will always be in
the correct place (with regard to character insertion/deletion), the cursor
printed to the screen will be incorrect when displaying wide Unicode like `안`.
When using these special characters the displayed cursor would need to be moved
according to the displayed width of the character. This display width depends on
the font used by the user. Correctly supporting the cursor with all Unicode will
thus stay out of scope for now.