Use tree-sitter to identify numbers - part 2
Computes number highlighting in the new `syntax` module. Limitations
due to the structure of `ERow` start showing. Source and highlighting
data should probably live in two separate vectors instead. Potentially,
the `syntax` module could be tasked with creating the `Highlight` vector.

`EditorSyntax` will also need to be updated to rules for highlighting
other tokens as well as choosing the rules for the correct language.