Use tree-sitter to identify numbers - part 3
Updates syntax for every changes. To work around multiple references to mutable
objects, the rows indices are passed to `editor_update_syntax` instead of the
mutable row directly. This makes for slightly more verbose code but also less
coupling between edition and syntax. This can hopefully serve as foundation when
moving from recomputation of the whole syntax tree (which is inefficient) to
update of the tree (as tree-sitter is intended to be used).