Use tree-sitter to identify numbers - part 4
Moves most of the syntax processing in the `syntax` module. Adds support for
syntax highlighting for all supported language (instead of always tring to parse
as rust code).

This is a "good enough" solution:

+ It uses tree-sitter and its queries to detect position of tokens to be
  highligted
+ It allows the addition of new languages
+ It supports cumstom highlighting for each language
+ The `syntax` module contains most of the logic related to parsing and
  highlighting
- The `EditorConfig::hl` field is still defined outside of the `syntax` module,
  with that we may be able to compute highlights for each token concurrently
- The definition of the highlitings rules is quite verbose

tree-sitter-highlight might provide a cleaner interface to highlighting and a
more scalable way to support new languages out of the box. It also provides with
helpers around language detection.

Note: A shortcut is used when detecting file types by encoding file extensions
directly in the code (instead of having them defined in a global array), this
should disappear when moving to tree-sitter-highlight.